requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
boto3==1.34.88
botocore==1.34.88
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==0.5.0
Django==4.2.2
django-cors-headers==4.3.1
django-storages==1.10.1
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.4
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.12
whitenoise==6.6.0
